+++ swarm/CHANGELOG.md
@@ -6,6 +6,8 @@
   There is a `Swarm`-scoped configuration for this version available since
   [1858](https://github.com/libp2p/rust-libp2p/pull/1858).
 
+- Update dependencies.
+
 # 0.27.2 [2021-02-04]
 
 - Have `ToggleProtoHandler` ignore listen upgrade errors when disabled.